Modifiada F05 para las dos versiones en cypress

+++ cypress.config.js
@@ -7,15 +7,20 @@ module.exports = {
         screenshotsFolder: `./${config.Results}/cypress/${config.Port}`,
         setupNodeEvents(on, config) {
             on('task', {
-                clearScreenshots(folder) {
-                    //const folder = `./${config.Results}/cypress/${config.Port}`;
+                clearScreenshots(datos) {
+                    const folder = datos[0];
+                    const pattern = datos[1];
                     if (fs.existsSync(folder)) {
                         const files = fs.readdirSync(folder);
+                        var del = 0;
                         files.forEach((file) => {
-                            const filePath = path.join(folder, file);
-                            fs.unlinkSync(filePath);
+                            if(file.indexOf(pattern) !== -1){
+                                const filePath = path.join(folder, file);
+                                fs.unlinkSync(filePath);
+                                del++;
+                            }
                         });
-                        return `Deleted ${files.length} files.`;
+                        return `Deleted ${del} files de ${pattern}.`;
                     }
                     return `Screenshots folder "${folder} "does not exist.`;
                 },
